requirements.txt
hydra-core==1.3.2
imgaug==0.4.0
numpy==1.23.4
omegaconf==2.2.3
opencv_python==4.6.0.66
Pillow==9.3.0
Pillow==10.1.0
PyYAML==6.0
PyYAML==6.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
webcolors==1.12
